+++ core/perf_test/PerfTest_ViewFill.hpp
@@ -126,7 +126,7 @@ static void ViewFill_Raw(benchmark::State& state) {
 }
 
 [[maybe_unused]] static void ViewFill_Rank1Strided(benchmark::State& state) {
-  const int N8 = std::pow(state.range(0), 8);
+  const size_t N8 = std::pow(state.range(0), 8);
 
   // allocate 2x the size since layout only has 1/2 the elements
   Kokkos::View<double*> a("A1", N8 * 2);
